Extract Table plotter into its own module
plots.py had grown large. Move TablePlotter and its _FormattedTablePlot
helper (and the hv.Store registration) into a focused table_plotter
module, importing the shared base and helpers from plots, mirroring how
flatten_plotter is structured.
